acknowledge the existence of a body
